+++ crates/holochain_types/CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## \[Unreleased\]
 
+- InstalledAppCommon field `role_assignments` is now an IndexMap to ensure consisent ordering.
+
 ## 0.5.0-dev.2
 
 ## 0.5.0-dev.1